requirements.txt
absl-py==0.11.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
importlib-metadata==3.7.0
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.20.1
oauthlib==3.1.0
Pillow==8.1.0
protobuf==3.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.0
